all test passing

+++ database/migrations/task1/2021_11_08_092943_create_comments_table.php
@@ -13,15 +13,17 @@ class CreateCommentsTable extends Migration
      */
     public function up()
     {
-        Schema::create('comments', function (Blueprint $table) {
-            $table->id();
-            $table->unsignedInteger('user_id');
-            $table->foreign('user_id')->references('id')->on('users');
-            $table->unsignedInteger('comment_id');
-            $table->foreign('comment_id')->references('id')->on('comments');
-            $table->string('comment_text');
-            $table->timestamps();
-        });
+        Schema::create(
+            'comments', function (Blueprint $table) {
+                $table->id();
+                $table->unsignedBigInteger('user_id');
+                $table->foreign('user_id')->references('id')->on('users');
+                $table->unsignedBigInteger('comment_id');
+                $table->foreign('comment_id')->references('id')->on('comments');
+                $table->string('comment_text');
+                $table->timestamps();
+            }
+        );
     }
 
     /**
